Move calculate_classified_track_density tests to test_track_density.py
Also cover the disjoint-grid path: all-NaN densities collapse to the
uniform empty result shape in both sum and UD modes.
